+++ frontend/src/routes/router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router";
 
-import App from "./App";
+import App from "../App";
 
 const router = createBrowserRouter([
     {
